KSH rework
Added a command dostuff to print some stuff.

Changed the signature of ksh_get_line_number_from_bottom to return a signed int because it needs to check if the line to render is under 0.

+++ ksh.h
@@ -42,7 +42,7 @@ char* ksh_get_current_type_line();
 BOOL ksh_is_current_type_line(int nb);
 int ksh_get_current_line_nb();
 
-uint8_t ksh_get_line_number_from_bottom(uint8_t nb);
+int ksh_get_line_number_from_bottom(int nb);
 
 void ksh_push_lines();
